fatmapper: Fix multiplication overflow calculating FS size.

+++ fatmapper.c
@@ -425,11 +425,11 @@ int main(int argc, char *argv[])
 				sqlite3_errstr(wf.wf_db_err));
 		goto out;
 	}
-	total_bytes = fs->clusters * fs->cluster_size;
+	total_bytes = (uint64_t)fs->clusters * fs->cluster_size;
 
 	collect_fs_stats(&wf.base, (char *)fsdev, fs->cluster_size,
 			fs->cluster_size, total_bytes,
-			fs->free_clusters * fs->cluster_size,
+			(uint64_t)fs->free_clusters * fs->cluster_size,
 			0, 0, FAT_MAX_NAME_LEN);
 	CHECK_ERROR("while storing fs stats");
